Test the support page
The support page looks a bit different depending on whether you're
signed in or not, so I added a class for each case. Both classes rely
on a common class which tests all of the functionality that is the same
whether you're signed in or not. I considered making each class *extend*
the common class, but it would've made things too awkward.